Migrate typedefs to types.h file
