+++ software/frequency_relay/frequency_relay.h
@@ -1,5 +1,9 @@
+// frequency_relay.h
+
 #ifndef __FREQUENCY_RELAY_H_
 #define __FREQUENCY_RELAY_H_
+
+#include "types.h"
 
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
@@ -11,40 +15,6 @@
 #define FREQDATA_Q_LENGTH 20
 #define NUM_LOADS 5
 #define TIMING_LOG_SIZE 5
-
-typedef struct {
-	float frequency;		// in Hz
-	float roc; 				// rate of change of frequency in Hz/s
-	unsigned int n;			// raw ADC sample count
-	TickType_t timestamp;	// FreeRTOS tick count when measured
-} freqData_t;
-
-typedef enum {
-	LOAD_OFF,		// switch is off
-	LOAD_ON,		// switch is on
-	LOAD_SHED		// relay has shed this load
-} loadStatus_t;
-
-typedef enum {
-	SYSTEM_NORMAL,			
-	SYSTEM_MAINTENANCE,		
-	SYSTEM_MANAGING		
-} systemState_t;
-
-typedef struct {
-	TickType_t recent[TIMING_LOG_SIZE];	// last 5 measurements
-	TickType_t minTime;
-	TickType_t maxTime;
-	TickType_t avgTime;
-	int count;								// total measurements taken
-} timingLog_t;
-
-typedef struct {
-	float TF_threshold;
-	float TROC_threshold;
-	enum Threshold threshold_edit_mode;
-	systemState_t system_state;
-} system_status_t;
 
 extern freqData_t freq_data;
 extern loadStatus_t load_status[NUM_LOADS];
