+++ software/frequency_relay/testfau.c
@@ -1,4 +1,4 @@
-// delete this file before submissionnnnn
+// testfau.c
 
 #include "frequency_relay.h"
 #include "freertos/FreeRTOS.h"
