+++ software/frequency_relay/vga.c
@@ -1,102 +1,104 @@
-void vga_display_task(void *pvParameters) {
-	// grab char buffer device handle from pv params
-	alt_up_char_buffer_dev *char_buffer = (alt_up_char_buffer_dev *)pvParameters;
+    // vga.c
 
-	// ~60Hz (17ms) timing setup
-	TickType_t xLastWakeTime = xTaskGetTickCount();
-	const TickType_t xFrequency = pdMS_TO_TICKS(17); // i DONT THINK FREQ IS ACCURATE
+    void vga_display_task(void *pvParameters) {
+        // grab char buffer device handle from pv params
+        alt_up_char_buffer_dev *char_buffer = (alt_up_char_buffer_dev *)pvParameters;
 
-	freqData_t local_freq_data;
-	system_status_t local_system_status;
-	int kbd_rx;
-	int ignore_next_key = 0; // track releases
-	char text_buffer[64];
+        // ~60Hz (17ms) timing setup
+        TickType_t xLastWakeTime = xTaskGetTickCount();
+        const TickType_t xFrequency = pdMS_TO_TICKS(17); // i DONT THINK FREQ IS ACCURATE
 
-	printf("VGA Task Started\n");
-	fflush(stdout);
+        freqData_t local_freq_data;
+        system_status_t local_system_status;
+        int kbd_rx;
+        int ignore_next_key = 0; // track releases
+        char text_buffer[64];
 
-	while(1) {
-		// -- process keyboard inputs --
-		while (xQueueReceive(kbdQ, &kbd_rx, 0) == pdTRUE) {
+        printf("VGA Task Started\n");
+        fflush(stdout);
 
-			// check for if its a release code
-			if (kbd_rx == PS2_BREAK) {
-				ignore_next_key = 1;
-				continue;
-			}
-			// break sequence means code is repeated after PS2_BREAK
-			if (ignore_next_key) {
-				ignore_next_key = 0;
-				continue;
-			}
+        while(1) {
+            // -- process keyboard inputs --
+            while (xQueueReceive(kbdQ, &kbd_rx, 0) == pdTRUE) {
 
-			// -- lock system status to change struct members --
-			if (xSemaphoreTake(systemStatusMutex, 0) == pdTRUE) {
-				switch (kbd_rx) {
-					case PS2_1:
-						system_status.threshold_edit_mode = TF;
-						break;
-					case PS2_2:
-						system_status.threshold_edit_mode = TROC;
-						break;
-					case PS2_UP:
-						if (system_status.threshold_edit_mode) {
-							system_status.TROC_threshold += 0.5;
-						} else {
-							system_status.TF_threshold += 0.5;
-						}
-						break;
-					case PS2_DOWN:
-						if (system_status.threshold_edit_mode) {
-							system_status.TROC_threshold -= 0.5;
-						} else {
-							system_status.TF_threshold -= 0.5;
-						}
-						break;
-				}
-				// -- unlock semaphore --
-				xSemaphoreGive(systemStatusMutex);
-			}
-		}
+                // check for if its a release code
+                if (kbd_rx == PS2_BREAK) {
+                    ignore_next_key = 1;
+                    continue;
+                }
+                // break sequence means code is repeated after PS2_BREAK
+                if (ignore_next_key) {
+                    ignore_next_key = 0;
+                    continue;
+                }
 
-		// -- gather local copies of data for display --
-		// peak (don't steal from load management)
-		if (xQueuePeek(freqDataQ, &local_freq_data, 0) != pdTRUE) {
-			local_freq_data.frequency = 0;
-			local_freq_data.roc = 0;
-		}
+                // -- lock system status to change struct members --
+                if (xSemaphoreTake(systemStatusMutex, 0) == pdTRUE) {
+                    switch (kbd_rx) {
+                        case PS2_1:
+                            system_status.threshold_edit_mode = TF;
+                            break;
+                        case PS2_2:
+                            system_status.threshold_edit_mode = TROC;
+                            break;
+                        case PS2_UP:
+                            if (system_status.threshold_edit_mode) {
+                                system_status.TROC_threshold += 0.5;
+                            } else {
+                                system_status.TF_threshold += 0.5;
+                            }
+                            break;
+                        case PS2_DOWN:
+                            if (system_status.threshold_edit_mode) {
+                                system_status.TROC_threshold -= 0.5;
+                            } else {
+                                system_status.TF_threshold -= 0.5;
+                            }
+                            break;
+                    }
+                    // -- unlock semaphore --
+                    xSemaphoreGive(systemStatusMutex);
+                }
+            }
 
-		// lock then grab
-		if (xSemaphoreTake(systemStatusMutex, portMAX_DELAY)) {
-			local_system_status = system_status;
-			xSemaphoreGive(systemStatusMutex);
-		}
-		if (xSemaphoreTake(loadStatusMutex, portMAX_DELAY)) {
-			// TODO: local_load_status =
-			xSemaphoreGive(loadStatusMutex);
-		}
-		if (xSemaphoreTake(timingLogMutex, portMAX_DELAY)) {
-			// TODO: local_timing_log = timing_log;
-			xSemaphoreGive(timingLogMutex);
-		}
+            // -- gather local copies of data for display --
+            // peak (don't steal from load management)
+            if (xQueuePeek(freqDataQ, &local_freq_data, 0) != pdTRUE) {
+                local_freq_data.frequency = 0;
+                local_freq_data.roc = 0;
+            }
 
-		// -- render VGA elements --
-		sprintf(text_buffer, "[%d] TF Threshold: %5.2f Hz    ", 
-			(1 - system_status.threshold_edit_mode), local_system_status.TF_threshold);
-		alt_up_char_buffer_string(char_buffer, text_buffer, 5, 5);
-		
-		sprintf(text_buffer, "[%d] TROC Threshold: %5.2f Hz/s    ", 
-			system_status.threshold_edit_mode, local_system_status.TROC_threshold);
-		alt_up_char_buffer_string(char_buffer, text_buffer, 5, 7);
-		
-		sprintf(text_buffer, "Current Frequency: %f Hz    ", local_freq_data.frequency);
-		alt_up_char_buffer_string(char_buffer, text_buffer, 5, 9);
+            // lock then grab
+            if (xSemaphoreTake(systemStatusMutex, portMAX_DELAY)) {
+                local_system_status = system_status;
+                xSemaphoreGive(systemStatusMutex);
+            }
+            if (xSemaphoreTake(loadStatusMutex, portMAX_DELAY)) {
+                // TODO: local_load_status =
+                xSemaphoreGive(loadStatusMutex);
+            }
+            if (xSemaphoreTake(timingLogMutex, portMAX_DELAY)) {
+                // TODO: local_timing_log = timing_log;
+                xSemaphoreGive(timingLogMutex);
+            }
 
-		sprintf(text_buffer, "Current RoC: %f Hz/s    ", local_freq_data.roc);
-		alt_up_char_buffer_string(char_buffer, text_buffer, 5, 11);
+            // -- render VGA elements --
+            sprintf(text_buffer, "[%d] TF Threshold: %5.2f Hz    ", 
+                (1 - system_status.threshold_edit_mode), local_system_status.TF_threshold);
+            alt_up_char_buffer_string(char_buffer, text_buffer, 5, 5);
+            
+            sprintf(text_buffer, "[%d] TROC Threshold: %5.2f Hz/s    ", 
+                system_status.threshold_edit_mode, local_system_status.TROC_threshold);
+            alt_up_char_buffer_string(char_buffer, text_buffer, 5, 7);
+            
+            sprintf(text_buffer, "Current Frequency: %f Hz    ", local_freq_data.frequency);
+            alt_up_char_buffer_string(char_buffer, text_buffer, 5, 9);
 
-		// TODO: add load and timing stats here
+            sprintf(text_buffer, "Current RoC: %f Hz/s    ", local_freq_data.roc);
+            alt_up_char_buffer_string(char_buffer, text_buffer, 5, 11);
 
-		vTaskDelayUntil(&xLastWakeTime, xFrequency);
-	}
-}
+            // TODO: add load and timing stats here
+
+            vTaskDelayUntil(&xLastWakeTime, xFrequency);
+        }
+    }
